upgrade b4.5

+++ MeadowExpressivePixels/MeadowApp.cs
@@ -82,7 +82,7 @@ namespace MeadowExpressivePixels
                 chipSelectPin: null,
                 dcPin: Device.Pins.D01,
                 resetPin: Device.Pins.D00,
-                width: (uint)displayWidth, height: (uint)displayHeight);
+                width: displayWidth, height: displayHeight);
 
             graphics = new GraphicsLibrary(display);
